Fix circular dependency by moving _get_cell_from_lattice_p function to utils.
